refactor: move tenant.Service unit tests into its package
Delete the top-level TenantService interface now that it's not used.
