--- tenant.go
@@ -1,10 +1,0 @@
-package influxdb
-
-// TenantService is a service that exposes the functionality of the embedded services.
-type TenantService interface {
-	UserService
-	PasswordsService
-	UserResourceMappingService
-	OrganizationService
-	BucketService
-}
